+++ src/main/java/com/petmatz/PetmetzApplication.java
@@ -1,0 +1,13 @@
+package com.petmatz;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PetmetzApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PetmetzApplication.class, args);
+	}
+
+}
