+++ src/main/java/com/petmatz/api/example/TestDTO.java
@@ -1,0 +1,11 @@
+package com.petmatz.api.example;
+
+import lombok.Data;
+
+@Data
+public class TestDTO {
+
+    private String name;
+    private String age;
+
+}
