+++ src/main/java/com/petmatz/common/exception/BaseErrorCode.java
@@ -1,0 +1,5 @@
+package com.petmatz.common.exception;
+
+public interface BaseErrorCode {
+    ErrorReason getErrorReason();
+}
