+++ src/main/java/com/petmatz/common/exception/DomainException.java
@@ -1,0 +1,15 @@
+package com.petmatz.common.exception;
+
+import lombok.Getter;
+
+@Getter
+public class DomainException extends CustomException {
+
+    public DomainException(BaseErrorCode errorCode, String message) {
+        super(errorCode, message);
+    }
+
+    public DomainException(BaseErrorCode errorCode) {
+        super(errorCode, "도메인 계층 예외");
+    }
+}
