+++ src/main/java/com/petmatz/common/exception/ErrorReason.java
@@ -1,0 +1,7 @@
+package com.petmatz.common.exception;
+
+public record ErrorReason(Integer status, String errorCode, String message) {
+    public static ErrorReason of(Integer status, String errorCode, String message) {
+        return new ErrorReason(status, errorCode, message);
+    }
+}
