+++ src/main/java/com/petmatz/common/exception/WebException.java
@@ -1,0 +1,11 @@
+package com.petmatz.common.exception;
+
+import lombok.Getter;
+
+@Getter
+public class WebException extends CustomException {
+
+    public WebException(BaseErrorCode errorCode) {
+        super(errorCode, "웹 계층 예외");
+    }
+}
