+++ src/main/java/com/petmatz/domain/example/Test.java
@@ -1,0 +1,28 @@
+package com.petmatz.domain.example;//package spring.basic.domain.test;
+//
+//import jakarta.persistence.Entity;
+//import jakarta.persistence.GeneratedValue;
+//import jakarta.persistence.GenerationType;
+//import jakarta.persistence.Id;
+//import lombok.AccessLevel;
+//import lombok.Builder;
+//import lombok.Getter;
+//import lombok.NoArgsConstructor;
+//import spring.basic.domain.global.BaseEntity;
+//
+//@Entity
+//@NoArgsConstructor(access = AccessLevel.PROTECTED)
+//@Getter
+//public class Test extends BaseEntity {
+//
+//    @Id
+//    @GeneratedValue(strategy = GenerationType.IDENTITY)
+//    private Long id;
+//
+//    private String name;
+//
+//    @Builder
+//    public Test(String name) {
+//        this.name = name;
+//    }
+//}
