+++ src/main/java/com/petmatz/domain/example/TestReader.java
@@ -1,0 +1,16 @@
+package com.petmatz.domain.example;//package spring.basic.domain.test;
+//
+//import lombok.RequiredArgsConstructor;
+//import org.springframework.stereotype.Component;
+//
+//@Component
+//@RequiredArgsConstructor
+//public class TestReader {
+//
+//    private final TestRepository repository;
+//
+//    public void getTest() {
+//        repository.findAll();
+//    }
+//
+//}
