+++ src/main/java/com/petmatz/domain/example/TestRepository.java
@@ -1,0 +1,6 @@
+package com.petmatz.domain.example;//package spring.basic.domain.test;
+//
+//import org.springframework.data.jpa.repository.JpaRepository;
+//
+//public interface TestRepository extends JpaRepository<Test, Long> {
+//}
