+++ src/main/java/com/petmatz/domain/example/TestService.java
@@ -1,0 +1,21 @@
+package com.petmatz.domain.example;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import spring.basic.common.exception.DomainException;
+import spring.basic.common.exception.GlobalErrorCode;
+import spring.basic.common.exception.WebException;
+
+@Service
+@RequiredArgsConstructor
+public class TestService {
+
+    private final TestOpenApiInterface testOpenApiInterface;
+//    private final TestReader testReader;
+
+    public void testGet() {
+        testOpenApiInterface.getTest();
+//        throw new DomainException(GlobalErrorCode.PERMISSION_DENIED);
+//        testReader.getTest();
+    }
+}
