+++ src/main/java/com/petmatz/infra/example/TestOpenApi.java
@@ -1,0 +1,13 @@
+package com.petmatz.infra.example;
+
+import org.springframework.stereotype.Component;
+
+import spring.basic.domain.test.TestOpenApiInterface;
+
+@Component
+public class TestOpenApi implements TestOpenApiInterface {
+    @Override
+    public void getTest() {
+
+    }
+}
