+++ src/main/java/com/petmatz/infra/websocket/WebSocketConfig.java
@@ -1,0 +1,27 @@
+package com.petmatz.infra.websocket;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.messaging.simp.config.MessageBrokerRegistry;
+import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
+import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
+import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
+
+@Configuration
+@EnableWebSocketMessageBroker
+public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
+
+
+    @Override
+    public void configureMessageBroker(MessageBrokerRegistry registry) {
+        WebSocketMessageBrokerConfigurer.super.configureMessageBroker(registry);
+    }
+
+    @Override
+    public void registerStompEndpoints(StompEndpointRegistry registry) {
+        WebSocketMessageBrokerConfigurer.super.registerStompEndpoints(registry);
+    }
+
+
+
+
+}
