--- src/main/java/com/petmetz/PetmetzApplication.java
@@ -1,13 +1,0 @@
-package com.petmetz;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PetmetzApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PetmetzApplication.class, args);
-	}
-
-}
