--- src/main/java/com/petmetz/api/example/TestDTO.java
@@ -1,11 +1,0 @@
-package com.petmetz.api.example;
-
-import lombok.Data;
-
-@Data
-public class TestDTO {
-
-    private String name;
-    private String age;
-
-}
