--- src/main/java/com/petmetz/common/config/ThreadLocalLogTrace.java
@@ -1,81 +1,0 @@
-package com.petmetz.common.config;
-
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Component;
-
-@Slf4j
-@Component
-public class ThreadLocalLogTrace {
-
-
-    private static final String START_PREFIX = "-->";
-    private static final String COMPLETE_PREFIX = "<--";
-    private static final String EX_PREFIX = "<X-";
-
-    private ThreadLocal<TraceId> traceIdHolder = new ThreadLocal<>(); // 동시성 이슈 발생, traceId 동기화
-
-    public TraceStatus begin(String message) {
-        syncTraceId();
-        TraceId traceId = traceIdHolder.get();
-        Long startTimeMs = System.currentTimeMillis();
-        log.info("[{}] {}{}", traceId.getId(), addSpace(START_PREFIX,
-                traceId.getLevel()), message);
-        return new TraceStatus(traceId, startTimeMs, message);
-    }
-
-    private void syncTraceId() {
-        TraceId traceId = traceIdHolder.get();
-        if (traceId == null) {
-            traceIdHolder.set(new TraceId());
-        } else {
-            traceIdHolder.set(traceId.createdNextId());
-        }
-    }
-
-
-    public void end(TraceStatus status) {
-        complete(status, null);
-    }
-
-    public void exception(TraceStatus status, Exception e) {
-        complete(status, e);
-    }
-
-    private void complete(TraceStatus status, Exception e) {
-        Long stopTimeMs = System.currentTimeMillis();
-        long resultTimeMs = stopTimeMs - status.getStartTimeMs();
-        TraceId traceId = status.getTraceId();
-        if (e == null) {
-            log.info("[{}] {}{} time={}ms", traceId.getId(),
-                    addSpace(COMPLETE_PREFIX, traceId.getLevel()), status.getMessage(),
-                    resultTimeMs);
-        } else {
-            log.info("[{}] {}{} time={}ms ex={}", traceId.getId(),
-                    addSpace(EX_PREFIX, traceId.getLevel()), status.getMessage(), resultTimeMs,
-                    e.toString());
-        }
-
-        releaseTraceId();
-    }
-
-    private void releaseTraceId() {
-        TraceId traceId = traceIdHolder.get();
-        if (traceId.isFirstLevel()) {
-            traceIdHolder.remove(); // destroy
-        } else {
-            traceIdHolder.set(traceId.createdPreviousId());
-            ;
-        }
-    }
-
-    private static String addSpace(String prefix, int level) {
-        StringBuilder sb = new StringBuilder();
-        for (int i = 0; i < level; i++) {
-            sb.append((i == level - 1) ? "|" + prefix : "|   ");
-        }
-        return sb.toString();
-    }
-
-}
-
