--- src/main/java/com/petmetz/common/exception/BaseErrorCode.java
@@ -1,5 +1,0 @@
-package com.petmetz.common.exception;
-
-public interface BaseErrorCode {
-    ErrorReason getErrorReason();
-}
