--- src/main/java/com/petmetz/common/exception/CustomException.java
@@ -1,24 +1,0 @@
-package com.petmetz.common.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public abstract class CustomException extends RuntimeException {
-
-    protected final BaseErrorCode errorCode;
-    protected final String sourceLayer;
-
-    public Integer getStatus() {
-        return errorCode.getErrorReason().status();
-    }
-
-    public String getMessage() {
-        if (sourceLayer == null) {
-            return errorCode.getErrorReason().message();
-        } else {
-            return String.format("%s - %s", sourceLayer, errorCode.getErrorReason().message());
-        }
-    }
-}
