--- src/main/java/com/petmetz/common/exception/DomainException.java
@@ -1,15 +1,0 @@
-package com.petmetz.common.exception;
-
-import lombok.Getter;
-
-@Getter
-public class DomainException extends CustomException {
-
-    public DomainException(BaseErrorCode errorCode, String message) {
-        super(errorCode, message);
-    }
-
-    public DomainException(BaseErrorCode errorCode) {
-        super(errorCode, "도메인 계층 예외");
-    }
-}
