--- src/main/java/com/petmetz/common/exception/ErrorReason.java
@@ -1,7 +1,0 @@
-package com.petmetz.common.exception;
-
-public record ErrorReason(Integer status, String errorCode, String message) {
-    public static ErrorReason of(Integer status, String errorCode, String message) {
-        return new ErrorReason(status, errorCode, message);
-    }
-}
