--- src/main/java/com/petmetz/common/exception/WebException.java
@@ -1,11 +1,0 @@
-package com.petmetz.common.exception;
-
-import lombok.Getter;
-
-@Getter
-public class WebException extends CustomException {
-
-    public WebException(BaseErrorCode errorCode) {
-        super(errorCode, "웹 계층 예외");
-    }
-}
