--- src/main/java/com/petmetz/domain/example/TestOpenApiInterface.java
@@ -1,7 +1,0 @@
-package com.petmetz.domain.example;
-
-public interface TestOpenApiInterface {
-
-    void getTest();
-
-}
