--- src/main/java/com/petmetz/domain/example/TestReader.java
@@ -1,16 +1,0 @@
-package com.petmetz.domain.example;//package spring.basic.domain.test;
-//
-//import lombok.RequiredArgsConstructor;
-//import org.springframework.stereotype.Component;
-//
-//@Component
-//@RequiredArgsConstructor
-//public class TestReader {
-//
-//    private final TestRepository repository;
-//
-//    public void getTest() {
-//        repository.findAll();
-//    }
-//
-//}
