--- src/main/java/com/petmetz/domain/example/TestRepository.java
@@ -1,6 +1,0 @@
-package com.petmetz.domain.example;//package spring.basic.domain.test;
-//
-//import org.springframework.data.jpa.repository.JpaRepository;
-//
-//public interface TestRepository extends JpaRepository<Test, Long> {
-//}
