--- src/main/java/com/petmetz/infra/example/TestOpenApi.java
@@ -1,13 +1,0 @@
-package com.petmetz.infra.example;
-
-import org.springframework.stereotype.Component;
-
-import spring.basic.domain.test.TestOpenApiInterface;
-
-@Component
-public class TestOpenApi implements TestOpenApiInterface {
-    @Override
-    public void getTest() {
-
-    }
-}
