+++ src/test/java/com/petmatz/PetmetzApplicationTests.java
@@ -1,0 +1,13 @@
+package com.petmatz;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PetmetzApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
