--- src/test/java/com/petmetz/PetmetzApplicationTests.java
@@ -1,13 +1,0 @@
-package com.petmetz;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PetmetzApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
